refactor(git): create `GitRunInfo::run_silent`
