+++ src/git/run.rs
@@ -83,4 +83,57 @@ impl GitRunInfo {
             .with_context(|| format!("Converting exit code {} from i32 to isize", exit_code))?;
         Ok(exit_code)
     }
+
+    /// Run Git silently (don't display output to the user).
+    ///
+    /// Whenever possible, use `git2`'s bindings to Git instead, as they're
+    /// considerably more lightweight and reliable.
+    ///
+    /// Returns the stdout of the Git invocation.
+    pub fn run_silent<S: AsRef<str> + std::fmt::Debug>(
+        &self,
+        repo: &git2::Repository,
+        event_tx_id: Option<EventTransactionId>,
+        args: &[S],
+    ) -> anyhow::Result<String> {
+        let GitRunInfo {
+            path_to_git,
+            working_directory,
+            env,
+        } = self;
+
+        // Technically speaking, we should be able to work with non-UTF-8 repository
+        // paths. Need to make the typechecker accept it.
+        let repo_path = repo.path();
+        let repo_path = repo_path.to_str().ok_or_else(|| {
+            anyhow::anyhow!(
+                "Path to Git repo could not be converted to UTF-8 string: {:?}",
+                repo_path
+            )
+        })?;
+
+        let args = {
+            let mut result = vec!["-C", repo_path];
+            result.extend(args.iter().map(|arg| arg.as_ref()));
+            result
+        };
+        let mut command = Command::new(path_to_git);
+        command.args(&args);
+        command.current_dir(working_directory);
+        command.env_clear();
+        command.envs(env.iter());
+        if let Some(event_tx_id) = event_tx_id {
+            command.env(BRANCHLESS_TRANSACTION_ID_ENV_VAR, event_tx_id.to_string());
+        }
+        let result = command
+            .output()
+            .with_context(|| format!("Spawning Git subprocess: {:?} {:?}", path_to_git, args))?;
+        let result = String::from_utf8(result.stdout).with_context(|| {
+            format!(
+                "Decoding stdout from Git subprocess: {:?} {:?}",
+                path_to_git, args
+            )
+        })?;
+        Ok(result)
+    }
 }
